change uint256 to uint

+++ contracts/SupplyChain.sol
@@ -6,10 +6,10 @@ contract SupplyChain {
     address public owner;
 
     // <skuCount>
-    uint256 public skuCount;
+    uint public skuCount;
 
     // <items mapping>
-    mapping(uint256 => Item) items;
+    mapping(uint => Item) items;
 
     // <enum State: ForSale, Sold, Shipped, Received>
     enum State {
@@ -22,8 +22,8 @@ contract SupplyChain {
     // <struct Item: name, sku, price, state, seller, and buyer>
     struct Item {
         string name;
-        uint256 sku;
-        uint256 price;
+        uint sku;
+        uint price;
         State state;
         address payable seller;
         address payable buyer;
@@ -34,16 +34,16 @@ contract SupplyChain {
      */
 
     // <LogForSale event: sku arg>
-    event LogForSale(uint256 sku);
+    event LogForSale(uint sku);
 
     // <LogSold event: sku arg>
-    event LogSold(uint256 sku);
+    event LogSold(uint sku);
 
     // <LogShipped event: sku arg>
-    event LogShipped(uint256 sku);
+    event LogShipped(uint sku);
 
     // <LogReceived event: sku arg>
-    event LogReceived(uint256 sku);
+    event LogReceived(uint sku);
 
     /*
      * Modifiers
@@ -62,16 +62,16 @@ contract SupplyChain {
         _;
     }
 
-    modifier paidEnough(uint256 _price) {
+    modifier paidEnough(uint _price) {
         require(msg.value >= _price, "Not enough Ether sent");
         _;
     }
 
-    modifier checkValue(uint256 _sku) {
+    modifier checkValue(uint _sku) {
         //refund them after pay for item (why it is before, _ checks for logic before func)
         _;
         Item memory item = items[_sku];
-        uint256 refund = msg.value - item.price;
+        uint refund = msg.value - item.price;
         (bool sent, bytes memory data) = item.buyer.call.value(refund)("");
         require(sent, "Refund could not be sent back");
     }
@@ -87,7 +87,7 @@ contract SupplyChain {
     // name, sku, price (we don't give free items), seller are non-zero
 
     // modifier forSale
-    modifier forSale(uint256 _sku) {
+    modifier forSale(uint _sku) {
         Item memory item = items[_sku];
         require(
             item.state == State.ForSale &&
@@ -99,19 +99,19 @@ contract SupplyChain {
     }
 
     // modifier sold(uint _sku)
-    modifier sold(uint256 _sku) {
+    modifier sold(uint _sku) {
         require(items[_sku].state == State.Sold, "Item already sold");
         _;
     }
 
     // modifier shipped(uint _sku)
-    modifier shipped(uint256 _sku) {
+    modifier shipped(uint _sku) {
         require(items[_sku].state == State.Shipped, "Item already shipped");
         _;
     }
 
     // modifier received(uint _sku)
-    modifier received(uint256 _sku) {
+    modifier received(uint _sku) {
         require(items[_sku].state == State.Received, "Item already received");
         _;
     }
@@ -123,12 +123,12 @@ contract SupplyChain {
         // skuCount is uint and starts with 0
     }
 
-    function getItem(uint256 sku)
+    function getItem(uint sku)
         external
         view
         returns (
             string memory name,
-            uint256 price,
+            uint price,
             address seller
         )
     {
@@ -136,7 +136,7 @@ contract SupplyChain {
         return (item.name, item.price, item.seller);
     }
 
-    function addItem(string memory _name, uint256 _price)
+    function addItem(string memory _name, uint _price)
         public
         returns (bool)
     {
@@ -171,7 +171,7 @@ contract SupplyChain {
     //    - check the value after the function is called to make
     //      sure the buyer is refunded any excess ether sent.
     // 6. call the event associated with this function!
-    function buyItem(uint256 sku)
+    function buyItem(uint sku)
         public
         payable
         forSale(sku)
@@ -194,7 +194,7 @@ contract SupplyChain {
     //    - the person calling this function is the seller.
     // 2. Change the state of the item to shipped.
     // 3. call the event associated with this function!
-    function shipItem(uint256 sku)
+    function shipItem(uint sku)
         public
         sold(sku)
         verifyCaller(items[sku].seller)
@@ -208,7 +208,7 @@ contract SupplyChain {
     //    - the person calling this function is the buyer.
     // 2. Change the state of the item to received.
     // 3. Call the event associated with this function!
-    function receiveItem(uint256 sku)
+    function receiveItem(uint sku)
         public
         shipped(sku)
         verifyCaller(items[sku].buyer)
@@ -218,14 +218,14 @@ contract SupplyChain {
     }
 
     // Uncomment the following code block. it is needed to run tests
-    function fetchItem(uint256 _sku)
+    function fetchItem(uint _sku)
         public
         view
         returns (
             string memory name,
-            uint256 sku,
-            uint256 price,
-            uint256 state,
+            uint sku,
+            uint price,
+            uint state,
             address seller,
             address buyer
         )
@@ -233,7 +233,7 @@ contract SupplyChain {
         name = items[_sku].name;
         sku = items[_sku].sku;
         price = items[_sku].price;
-        state = uint256(items[_sku].state);
+        state = uint(items[_sku].state);
         seller = items[_sku].seller;
         buyer = items[_sku].buyer;
         return (name, sku, price, state, seller, buyer);
